The v1.1.1 Update
Hello there, today I have been working hard to fix bugs and make the
plugin more customizable.

Changes:
- Added Messages.yml, You can change some messages in there to what you
would like them to say.
- Added GUICustomization to the config.yml, This allows you to add none
clickable items into the GUI so that it can look nicer.
- New Default GUI look, I changed the default look so that you can see
the use of GUICustomization option in the config.
- Success Chance to the Black Scrolls in the Config.yml, You now can
control the range of the percents that you will get when you get a book
back from the black scrolls.
- Max Level for categories, You can control the cap level that a book
will be when coming out of a category.
- Readded the Info GUI, I have added a command to open the Info GUI. To
open it you do /CE Info and it will open a GUI with all the
enchantments and some info on each. I will be changing how it looks
later on so it looks nicer.​
Fixes:
- Black Scrolls, I have fixed black scrolls so that now they will work
properly.​