v1.3 Update
Hello there, I have added a highly requested scroll. White Scrolls and
Destroy rate. I have added things into the config.yml to allow a bit of
customizing for it and also added things to the categories for it.

Please either delete the old Config.yml and reload or go to the default
config on the main page and add all the new things to your config.yml.

Added:
- White Scrolls
- Destroy Rate
- Success and Destroy rate enabler/disabler
Changes:
- /Ce BlackScroll command is now /CE Scroll <Player> <Black/White>
<Amount>
- CrazyEnchantments.BlackScroll to CrazyEnchantments.Scroll