v1.4.3 Update || The More Compatibility Update
Hello there, resonantly I have been getting people report a lot of
errors that are the same. About 5 updates ago I added the Glowing
effect to the plugin giving Books and Items without vanilla
enchantments on them the Glowing effect. Well since I added that I have
been getting reports from servers that are not running the newest 1.8
and 1.9 versions. So in this update I have added support to the newest
1.7 version and all the 1.8 versions.