v1.4.4 Update
Hello there, I have fixed a bug where you could change the name of an
item to like OverLoad 20 and it would put the enchantment on the item
you drop that item on. I have also fixed the error when you do /CE Add
and the enchantment is null<Enchantment>. I have added to /CE Info on
the categories lores to each one to tell you what items can be
enchanted in that category.

Bug:
- Renaming items to <Enchantment> <Lvl> allowed you to put that
enchantment on an item.
- /CE Add made the enchantments say null<Enchantment>​
Added:
- /CE Info the categories now show the items that those enchantments
can be added to.​