v1.4.5 Update || Bug fix and Magical Dust
Hello there, I have been asked quite a bit to add Dust so that if your
success rate is really low or your destroy rate is really high there is
a way to make it so the players can bring the success rate up and the
destroy rate down. I have also added it so in the config you can change
the lore on the Enchantment books.

Look here if you need to see the changes.
https://github.com/kicjow/Crazy-Enchantments/blob/master/config.yml

Bug:
- Enchantments giving errors​
Added:
- Magical Success Dust: allows you to make your enchantment books
success rate go up.
- Magical Destroy Dust: allows you to make your enchantment books
destroy rate go down.
- Enchantment book lore can now be changed in the config.yml.​