v1.5.1 Update || Factions Support and Bug fixs
Hello there, I have been working on fixing those pesky bugs and also
adding support for factions. I have added it so that if the player is
in the same faction as you or is an ally then it the enchantments will
not do any damage to them.

Bugs:
- When you buy a Enchantment Book with xp and then use an xp bottle you
get from the Tinkerer it gives the xp you spent back.
- /ce book didn't work for console.
Added:
- Faction support. If you hit your fellow faction member or ally with
an enchantment it will not do that enchant effect on that player.