v1.6.1 Update || The Lost Book Update
Hello there, Last night I was working hard on adding what I call Lost
Books. Lost Books are books that are old and dirty. The description for
them is that they are said to contain a random enchantment from the
category on the book. So you just need to clean them (Right Click the
Book) a bam you may find your self a enchantment book. (You will)

Added:
- Lost Books. They can be obtained by either the new command /CE
LostBook <Category> [Amount] [Player] or by buying them in the /CE GUI.
- Better support for all 1.8 and 1.9 versions.
- New lines into the Config.yml and Messages.yml
Changes:
- Inquisitive is now more balanced. It will not happen 100% of the time
but instead 1/7 of a chance but for every power the enchantment goes up
it lowers that chance by 1. So power 2 is 1/6 chance, power 3 is 1/5
chance and so on.
- Experience is now gives a bit lower xp.
- Tried to balance how the destroy and success rates work.
Fixes:
- Auto Smelt now doesn't work in none breaking block zones.
- Telepathy now doesn't work in none breaking block zones.
- Experience now doesn't work in none breaking block zones.