v1.6.1.3 Update || Bug Fix
Hello this update fixes a small bug with the Black Smith. He was ill
and was so he didn't want to check if he was adding enchantments that
bypassed the Max Enchantment Limit. I gave him some medicine and he is
feeling better and is now checking the limit.

Bug:
- BlackSmith would add more enchantments then the max enchantments
allowed.