v1.6.2 Update || New Protection Crystal and More Enchantments.
Hello, In this update I have added 10 new enchantments. I have also
added a new Crystal. The Protection Crystal you place on any item. If
you are to die and a item has this on it it will come back to you on
respawn.

New:
- /CE Crystal/C [Amount] [Player] - Gives a player a Protection Crystal.
Changes:
- Big code cleanup removed around 80KBs off the files size.
- Nursery now only heals health and not food. It has also been nurffed.
- /CE Info now is 18 lines long and I have added a new Category called
Misc. Right now it holds only 1 enchantment.
Bugs:
- Trap did not work when hitting players.
- Insomnia has been fixed so it can do what is was originally suppose
to do. It gives Mining Fatigue and Confusion but has a 1/3 chance to do
2x damage.
- Tinkerer let you put a book and item in at the same time. Now is only
book on book and item on item.
New Enchantments: (All enchantments have been added to /CE Info)
- MultiArrow: (Bow) For each tier it adds 1 arrow to your shots and has
a 1/3 chance of happening.
- Cactus: (Armor) Has a 1/4 chance to do small damage based on the Tier
to your attacker.
- Recover: (Armor) (Only Tier 1 will work) If you kill a player you get
Absorption and Regen for a small amount of time.
- StormCaller: (Armor) has a 1/12 chance to strike lightning on your
attacker.
- Implants: (Helmets) While walking you have a 1/25 (-1 for each tier)
to give you 1/2 a chicken nugget.(Food Bar)
- Rocket: (Boots) when less then 4 full hearts you have a 1/8 (-1 for
each Tier) to get launched away from your attacker.
- HellForged: (All Tools, Weapons, and Armor) While walking each item
with this enchantment has a 1/12 to fix your item by 1(+1 for every
Tier) durability.

These only work if you have a factions plugin.
- Angel: For each Tier adds 1 block to the default radius of 5. It
gives players that are allies in the radius Regen.
- LeaderShip: For each Tier adds 1 block to the default radius of 5. If
you have allies in your radius it will give you an extra .5 attack for
each ally in your radius.
- Commander: For each Tier adds 1 block to the default radius of 5. All
allies in your radius get Haste.