v1.6.4 Update || The API update.
Hello everyone, I have been working the paste few days on recoding the
plugin to use the API the plugin has. This update is more for
developers then server owners. But it does have some bug fixes owners
will need. Yes there has been a api but it wasn't being updated and
couldn't do much with it. I have added all the new enchantments to it
and made some new methods for the API.

Bugs:
- When getting a lvl 8 enchantment it sets the Roman Numeral as VII and
not VIII
- Faction Plugins make errors in console if not supported.
- When making a new Sign Shop the price is set to 0 and not the real
price.
- When players without the permission CrazyEnchantments.Sign place a
sign it tells them they do not have permission to use that command.
- GUI when at max of a row but not enough to make another row the row
doesn't expand to let the next button show.
- Berserker math for the enchantment was done wrong.
- Oxygenate and haste enchantments not working right
Updated:
- API for developers.
Added:
- EnchantmentUseEvent fires when an enchantment is activated.
- DisarmerUseEvent fires when the disarmer enchantment is activated.
- AngelUseEvent fires when angel enchantment is activated.
- All enchantments enums into the CEnchantments class.
- All available enchantment item types enums to EnchantmentType class.
- More methods into the CrazyEnchantments class.
- A new class called CEBook that allows you to make enchantment books
with ez.