v1.6.4.1 Update
This is a update to v1.6.4. This fixes all the bugs I could find that
v1.6.4 had made because of the big recode. Most things broke in that
update but now are fixed with this.

In this update all CustomEnchantments have been disabled. This is due
to the new api not being compatible with them and making a lot of
problem. I will be enabling them when I can make them compatible with
the API.

Bugs:
- Black Smith Broke
- Adding Enchantments Broke
- Black Scroll not working
-
Added:
- addEnchantment method to the API.
- removeEnchantment method to the API.