Fix for Unenchantable Gold Swords/Axes + Leather
Added "GOLD_SWORD" to isWeapon() and isSword().
Added "GOLD_AXE" to isWeapon() and isAxe().
Changed first appearance of "DIAMOND_HELMET", "DIAMOND_CHESTPLATE", "DIAMOND_LEGGINGS", and "DIAMOND_BOOTS" to leather equivalents. I assume the intention was to list leather here but was instead simply copy-pasted but left unchanged.

Golden swords and axes couldn't be enchanted, which is a shame! I looked in code and it seems to be the problem here. Also noticed "DIAMOND_BOOTS" and the like appeared twice, probably because it was meant to have "LEATHER_BOOTS" and visa versa. These seem to be the only errors I found going through this great plugin! Hope this change makes it in!