Pre-Release for v1.6.5.1
Added:
Sign Break Message and Send Enchantment Message needs added into the
Messages.yml.
Lore to the Black Smith's no result item and added the customization in
the config.yml.

Bug:
Error in the GUI.
When using dust on a book it doesn't change the %Description%
placeholder.

Fixed:
Spelling errors in the Enchantments.yml.