v1.6.5.1 Update || Lots of bug fixes.
Hello, in this update I have fixed a lot of bugs so that you have less
problems.

Added:
- Sign Break Message and Send Enchantment Message needs added into the
Messages.yml.
- Lore to the Black Smith's no result item and added the customization
in the config.yml.

Bug:
- Error in the GUI.
- When using dust on a book it doesn't change the %Description%
placeholder.
- Missing armor and weapons to the enchantments.
- Lightning enchantments would break players by making them have to
relog.
- The lightning enchantment for bows would hurt the user.
- Hitting NPCs making errors in console.

Fixed:
- Spelling errors in the Enchantments.yml.
- If particles are disabled for Wings it still would cause lag for the
server because its still trying to spawn particles.

Other:
- In Minecraft 1.9 and lower a bug happens when switching armor into
your hotbar by hitting a number and that slot having a block in it the
armor is taken off but the potion effects stay.
- In Minecraft 1.10 and up this bug doesn't happen. If you wish to
avoid this bug please update to 1.10.

Support:
- Thanks to @Scrydan for helping fix a armor and weapon bug.