v1.6.5.5 Update || Quick file fix.
Hello, this update isn’t a needed update. This is more for new players
who get this. I added s’s at the end of Axe, Sword, Helmet, and Bow. If
you have that in your file for where you specify what items the
enchantment goes on please remove the s from those.

Bug:
  - Had s’s in some item settings.