v1.7.1.5 Update
Hello, I have been working on trying to fix some bugs and add some more
support to help with the plugins performance.

Added:
  - Spartan Anti Cheat support.
https://www.spigotmc.org/resources/25638/
  - Better protection support.
  - In-Game enchantment name support for the GKitz.
  - This update should help with mcmmo and CoreProtect to register the
block breaking.

Changes:
  - Allows better PvE support for servers that like the PvP deny flag
but want a player to still use the enchants on mobs in the region.
  - Changed the default vanilla enchantment names in the GKitz to the
In-Game enchantment names and not the Code names.

Bug Fixes:
  - Tool and pick axe enchants would make the item unbreakable.