v1.7.3 Update || Less Bugs and Item Enchantment Upgrading
Hello, I have been working the past 6 hours to fix some bugs and add a
highly requested feature. This update has added the Item Enchantment
upgrade feature that has been spamming me for a while.

Added:
  - Armor Enchantment Upgrading option. Dropping a higher level
enchantment book on an item with a lower level of that enchantment
    will have a chance of upgrading the enchantment level to the books
level. If success full then the enchantment will be upgraded.
    If destroy rate happens then there is an option to ether destroy
the item or destroy the enchantment on the item.
    If nether are successful then the book is the only thing that
breaks.

Bug Fixes:
  - Rage making an error with its cooldown.
  - Dust could be used on lost books.
  - Custom Enchanted books not having the glowing effect even with the
toggle set to true.
  - When enchants are added they get added under lores. They now go
above the lores.
  - Using in-game enchantment names in the gkit and then looking in the
viewer option wont show the enchantment.