v1.7.3.4 Update || Major Duplication Bug Fix
Hello, I have been getting reports about a dumping bug but could never
replicate it. I have found out with the help of Doit that combat tag
plugins will dup items when the player has an item in the tinkerer or
black smith and then being in combat and leaving the server.

Bug Fixes:
  - Major duplication bug from the tinkerer and black smith.
  - Added some support to MegaSkills xp bottles.