v1.7.4.1 Update || The Bug Squashing Update
Hello, in this update I have been working to fix a lot of the reported
bugs. This should fix most of the reports I have seen. I have also made
it so the potion effect enchantments now work together. If a higher
level effect gets added then the higher potion will be added. If the
higher one is removed then the next highest will take the effects place.

Bug Fixes:
  - Right clicking shulker boxes in 1.11+ with enchanted armor would
give you the effect without putting the armor on.
  - 1.10+ if you click a hotbar button while hovering over an item that
is equipped it doesn't remove the potion effects.
  - If you click a hotbar button while hovering over an empty armor
slot and the hotbar button has armor with an enchantment on it it
wouldn't give the effects.
  - Switching two pieces of armor with potion enchantments would remove
the effect and not re add it.
  - Fixed up the equipping of the armor. I should work smoother now.
  - Disabled Shulker Boxes and Chests from the enchantment Telepathy.
This is to protect the items inside of them.
  - Info GUI's name didn't change with the new option which could allow
players to take items from that inventory.
  - Fortune not working on Blast.
  - Mining lapis with a custom enchantment would give ink sacks and not
lapis.
  - Telepathy not giving the silk touch items.

Added:
  - Switching enchanted armor will now not remove the potion effects if
the player has the enchantment on another piece of armor.
  - Potion effects will now be upgraded to the highest enchantment
level based on the players armor.
  - Potion effects now work together. The highest level of the effect
based on all the enchantments will be the one put on the player.
  - Mystery dust can now be obtained from the /CE Dust command.
  - New API methods in the CrazyEnchantments class