v1.7.5 Update || New Transmog Scroll and a bunch of bug fixes.
Hello, in this update I have been working on adding a highly requested
scroll known as the Transmog Scroll. If you don’t know what that is
basicly its a scroll that will organize your enchantments by category.
So the highest rarity category’s go 1st. This is very useful to make
your items look much nicer. It also edits the items name add adds the
number of enchantments on it to the name.

Bug Fixes:
  - Kingdoms giving errors when a kingdom for a player isn't found.
  - Telepathy not giving xp for ores.
  - Protection Crystals can be protected.
  - Upgrading an enchantment and if destroy rate happens but the item
is protected the protection is ignored.
  - If upgrading the item and its set to destroy the item when destroy
rate is activated it would say the message for if only the enchantment
was removed.
  - When upgrading if you have the enchantment set to break when
destroy rate activates it would ignore the protection from a white
scroll.
  - Lost books would try and add them selfs to enchantable items when
dropped on the item.
  - In 1.9 down if you have an enchanted item on and a piece of armor
with the same enchantments and you use the hot bar keys to move items
it would remove the effects.
  - Adding an enchantment by using a book adds the enchantment under
the lore and not above it.

Added:
  - Protection Crystal now can have a chance of working when the player
dies.
  - These items can now be stacked on your cursor when adding them to
an item. Success/Destroy Dust, Protection Crystals, and White/Black
Scrolls. (Does not work in Creative)
  - Transmog Scrolls allow you to order your enchantments by rarity. It
also shows the amount of enchantments on the item.