v1.7.6 Update || New Scrambler, User Request Features, and Bug Fixes.
Hello, I have been working on adding a new feature called the
Scrambler. The scrambler re-rolls the enchantment books success and
destroy rates. It also has a built in CSGO Style re-rolling feature
which can be disabled if it causes lag on your server. This is great
for servers that don’t want to use dust to edit the chances. I have
also added a few user requested features. This also has a bunch of bug
fixes that have been reported to me.

Bug Fixes:
  - API method misspelled.
  - A small error from gkitz.
  - Fortune and blast not working together.
  - Silk Touch and Blast were not working together.
  - Blast dupping items.
  - Using a transmog scroll on an item with no custom name would set
the name to the id name.

Added:
  - New Scrambler. This allows you to re-roll your enchantment book's
success and destroy rates.
  - 1.9+ now supports Mob Eggs.
  - An option to disable wings when enemies are near by.
  - A Cap option for the enchantment levels.
  - Option to make black scrolls have a chance of working.

Changes:
  - Default white scroll and black scrolls lores have been shortened to
make it easier to read.
  - Re-done the enchantments.yml to make the info easier to understand
and better looking.
  - Changed some of the enchantment's max levels in default file to
make it more balanced.