v1.7.6.3.3 Update || Client Crash Bug Fix
Hello, this update is to fix the client crashing when using an anvil
with a custom enchantment on an item.

Bug Fix:
 - Putting custom enchanted items in an anvil crashes the clients game.

Draw Back:
 - To fix this bug I had to remove support for 1.11+’s glowing feature
when you add a custom enchantment to the item. If there is no vanilla
enchantment the item will now not glow. This will only be if the player
is on 1.11+ because 1.11+ changed how they see the glowing effect. Any
player using a minecraft version under 1.11 will see the glowing. This
does not effect the GUI’s tho. All players will be able to see the
Glowing in the GUI’s.