v1.7.6.4 Update || New Command, Bug Fixes, and More.
Hello, in this update I have been working to redo how the currencies
are handled so that in the future I can add more currency support. I
have also added a new useful command so that players that want books to
drop on the ground could do that. An example would be for a boss fight
so when the boss dies it will run the spawn command that will drop at
the location put into the command. I have also added a new permission
to stop the aura enchantments from effecting you. Mainly needed for
staff. I have also removed vault as a dependency.

Bug Fixes:
  - Closing the blacksmith with two things in it would only give back
one item instead of both.

Added:
  - Prefix in front of the sign break message.
  - New command /CE Spawn <Enchantment>
[Level:#/World:<World>/X:#/Y:#/Z:#].
  - New permission crazyenchantments.aurabypass will stop the aura
enchantments from effecting players with the permission.

Changes:
  - Recoded how the economy controller in the plugin works. This will
allow me to add more economies in the furture.

Removed:
  - Vault is no longer a dependency.