v1.7.6.5 Update || Lots of bug fixes.
Hello, in this update I have fixes a lot of bugs that have been
reported. This update is highly recommended because it contains a
duplication bug fix.

Bug Fixes:
  - Bow enchantments should work better together.
  - /Gkit didn't have a permission but now is hooked to the
crazyenchantments.access permission.
  - BlackSmith would give the books they combine a real enchantment
instead of the glowing. This would cause major problems if used in an
anvil.
  - FireworkAPI I use to stop the firework damage would cause lag on
servers with lower versions then 1.11.2.
  - Duplication glitch with blast and telepathy.
  - You could add the protection crystal to another protection crystal
if the one your adding it to doesn't have the glowing.
  - Left over vault code in the main class made the plugin still not
work if you didn't have vault.
  - Using telepathy on a spawner without silkspawners would still give
you a normal spawner.

Added:
  - Permissions for all admin commands. Use crazyenchantments.<Command>
for only a specific command and crazyenchantments.admin for all
commands.

Changes:
  - Spelling mistake. Haste was called hast.
  - Thanks to tyler489 for his spelling mistake find and pull request.
Boots were suppose to be boost
  - Some code clean up.