Hello, in this update I have fixed a lot of bugs that I have found and that have been reported.
Bug Fixes:
  - Removing vanilla enchantments with /CE remove would remove the
enchantment but say it couldn't find the enchantment.
  - Using telepathy and silk touch on a redstone ore would not give you
the redstone ore.
  - Breaking an ore with telepathy and silk touch would give you XP.
  - Lapis ore, Redstone ore, and Glowstone would duplicate with blast
and silk-touch.
  - Having enchanted armor in your player crafting slots then shift
clicking it out would activate the effects.
  - 1.9 down if you moved the armor on you with hot key and there was a
different armor piece type in that slot it would move it but keep the
effects.
  - When a player dies and leaves without respawning they lose any
items they had with the protection of the protection crystal. They now
get the items when they rejoin and respawn. But if the server reloads
or restarts the player will lose the items.

Added:
  - An option in the config to let the blast enchantment do 1 damage
for every block it breaks or only one damage all together.
  - An option to toggle the update checker in the config.yml.

Changes:
  - Cleaned up the package names so they follow java standards.
  - Moved the Clouds: option under the Wings: options.