v1.7.8 Update || Bug fixes and more
Hello, in this update I have worked hard to fix bugs, add more things
and a bit more. I hope you enjoy this long change log.

Bug Fixes:
  - Adding Ice Freeze to a bow would also add Freeze when adding
another enchantment to the bow.
  - Using blast and silk touch on a block would then uses the metadata
of that block on all the other blocks.
  - /CE reload does not save currently loaded gkit cool-down times for
players that just got a gkit and they didn't log out.
  - The enchantment doctor was not working at all.

Added:
  - Added more documentation to the plugin's methods.
  - Some support for DakataAntiCheat and the blast enchantment.
  - Work has been started on v1.8 and methods are being added but will
not effect in-game.
  - Auto Backup that backs up all player data. This includes GKit
cool-downs and souls(Soon to come).
  - Toggle option for the auto backup message that will show up in the
console.
  - The new custom event called BuyBookEvent for when a player buys a
book in the GUI or a sign.

Changes:
  - A bunch of code clean up.
  - The doctor enchantment has been nerfed from (2 + power) to (1 +
power). This is because it would heal 1 and 1/2 hearts every hit.
  - Fully re-coded the GKitz system to work better for other developers
and much easier to use in my plugin.

Removed:
  - UnNeed methods in the main CrazyEnchantments class.