v1.7.9 Update || Bugs Bugs Go Away
Hello, so it has been a while since the last update. I have been taking
down all the bug reports I have seen and have fixed just about all of
them. I have also worked on adding things that players have requested
and wanted to be added. There has also been some file changes so make
sure to add the changes.

Bug Fixes:
- Sign's now getting a random ID in the Signs.yml so that they don't
overlap another signs ID.
- Protection Crystal signs would not give the crystal but it would
still buy it.
- Error when using a dropper and having it shoot out armor.
- Skulls with a potion enchantment on them would not add the effects
when put on.
- Having the enchantment name and custom name different in the
CustomEnchantments.yml would make the enchantment not work and shoot
errors.
- Some plugins that stop the players from right clicking the armor on
like MythicMobs would stop the armor from going on but still add the
enchantment effects.
- Putting an item in the GUI that is outside of its size would error.
Now the item will just be ignored and will open the GUI without
erroring.
- Putting two different items with an enchantment on it in the Black
Smith would work and turn into the main item with the upgraded
enchantments.
- Adding an enchantment via book would add the enchantment but not
add the glowing effect.
- Getting hack messages from No Cheat Plus with the Blast Enchantment.
- Suppressed an error from MCUpdate.org.
- An error would show in the console when using the dust command and
trying to give the console dust.
- Having Experience and Silk Touch on a pickaxe would still give you
XP even though you get the ore back.

Added:
- Now doing /ci, /clear, or /clearinventory will remove the potion
effects of the enchantments that are on your armor. This is so the
effects don't stay when u clear inventory.
- Now able to use placeholders in the GUICustomization option in the
config. They can be used in the name or lore. %vault% %xp_level%
%xp_total% http://prntscr.com/f3aiya http://prntscr.com/f3aiiz
- Now with support for AAC. If you still get kicked for CE
enchantments please contact me with the enchantment and the kick reason.
- You can now use Min-Max in the /CE book and /CE spawn commands.
Example: "/CE book overload 1-5 1" and it will give you an Overload
book with a level from 1-5.
- You can now use categories in the /CE spawn command to spawn lost
books. Example: "/ce spawn common" and it will spawn a lost book at
your feet.
- New option in the config to allow world guard owners and/or members
to be able to fly in regions. This does not check groups but only
specific players that are added in the regions.
- New option in the config so when a player right-clicks an
enchantment table it will open the /CE GUI instead of the enchantment
table GUI.
- New option in the config to stop from losing the protection from a
protection crystal on death.
- Boom and Self Destruct enchantments now push the player back from
the explosion instead of just damaging them.

Changes:
- Self Destruct enchantment now uses the same explosion effect as
Boom. This helps other plugins out because it doesn't use a real
explosion.
- Boom and Self Destruct now remove items on the ground when they
explode.
- Self Destruct now looses all items on death. Unless the item is
protected with a crystal then it will stay on respawn.
- Moved the Player's data saving when they leave to async to avoid
TPS drops.
- Allies used to only spawn when the player is in a PVP area. They
now spawn when the player takes damage or does damage.