Fixed NPE and removed spam
Fixed cheking getInventory for null instead of getClickedInventory (the one being used).
Since the event gets fired each time a user clicks an inventory, the NEED_TO_USE_PLAYER_INVENTORY message was being spammed when using chests, shulkers, etc. Moved the message to the part of the code in which we have checked that the user is trying to do a reroll.